+++ packages/annotation/tests/tools/cleanSvg.test.tsx
@@ -1,100 +1,5 @@
 import { render, screen } from '@testing-library/react'
-import { roundPxDecimalPlaces, cleanTransform } from '../../src/tools/helpers'
 import { cleanSvg } from '../../src/tools'
-
-describe('roundPxDecimalPlaces', () => {
-    it('leaves standard strings alone', () => {
-        expect(roundPxDecimalPlaces('abc', 2)).toEqual('abc')
-        expect(roundPxDecimalPlaces('10', 2)).toEqual('10')
-    })
-
-    it('rounds long decimals', () => {
-        expect(roundPxDecimalPlaces('10.0000003px', 3)).toEqual('10px')
-        expect(roundPxDecimalPlaces('10.123123px', 3)).toEqual('10.123px')
-        expect(roundPxDecimalPlaces('10.1238px', 3)).toEqual('10.124px')
-    })
-})
-
-describe('cleanTransform', () => {
-    it('handles undefined and null', () => {
-        expect(cleanTransform(undefined, 2)).toBeUndefined()
-    })
-
-    it('extract translateX and translateY with semicolon', () => {
-        const input = 'translateX(140px) translateY(20px);'
-        const expected = 'translate(140,20)'
-        expect(cleanTransform(input, 2)).toEqual(expected)
-    })
-
-    it('extract translateX and translateY without semicolon', () => {
-        const input = 'translateX(140px) translateY(20px)'
-        const expected = 'translate(140,20)'
-        expect(cleanTransform(input, 2)).toEqual(expected)
-    })
-
-    it('extract transform and apply rounding', () => {
-        const input = 'translateX(140.1111px) translateY(20.1111px)'
-        const expected = 'translate(140.11,20.11)'
-        expect(cleanTransform(input, 2)).toEqual(expected)
-    })
-
-    it('extract translateX without translateY, and vice versa', () => {
-        const inputX = 'translateX(140.1111px)'
-        const expectedX = 'translate(140.11,0)'
-        expect(cleanTransform(inputX, 2)).toEqual(expectedX)
-        const inputY = 'translateY(20.1111px)'
-        const expectedY = 'translate(0,20.11)'
-        expect(cleanTransform(inputY, 2)).toEqual(expectedY)
-    })
-
-    it('extract scaleX and scaleY', () => {
-        const input = 'scaleX(2) scaleY(3)'
-        const expected = 'scale(2,3)'
-        expect(cleanTransform(input, 2)).toEqual(expected)
-    })
-
-    it('extract scaleX and fill in scaleY, and vice versa', () => {
-        const inputX = 'scaleX(2)'
-        const expectedX = 'scale(2,1)'
-        expect(cleanTransform(inputX, 2)).toEqual(expectedX)
-        const inputY = 'scaleY(2)'
-        const expectedY = 'scale(1,2)'
-        expect(cleanTransform(inputY, 2)).toEqual(expectedY)
-    })
-
-    it('extract rotate', () => {
-        const input = 'rotate(10deg)'
-        const expected = 'rotate(10)'
-        expect(cleanTransform(input, 2)).toEqual(expected)
-    })
-
-    it('omit zero rotation', () => {
-        const input = 'translateX(10) rotate(0deg)'
-        const expected = 'translate(10,0)'
-        expect(cleanTransform(input, 2)).toEqual(expected)
-    })
-
-    it('detects no transform', () => {
-        const input = 'none'
-        expect(cleanTransform(input, 2)).toBeNull()
-    })
-
-    it('detects no transform (prefix)', () => {
-        const input = 'transform: none'
-        expect(cleanTransform(input, 2)).toBeNull()
-    })
-
-    it('leaves rotations', () => {
-        const input = 'rotate(90)'
-        expect(cleanTransform(input, 2)).toEqual(input)
-    })
-
-    it('extract skewX', () => {
-        const input = 'skewX(1.0002)'
-        const expected = 'skewX(1.0002)'
-        expect(cleanTransform(input, 2)).toEqual(expected)
-    })
-})
 
 describe('cleanSvg', () => {
     it('rounds values in one element', () => {
@@ -233,7 +138,6 @@ describe('cleanSvg', () => {
     })
 
     it('removes nested g with role dimensions-reference', () => {
-        const transform = 'none'
         render(
             <svg role={'svg'}>
                 <g role={'dimensions-reference'}>